requirements.txt
altgraph==0.17.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
gevent==23.9.1
greenlet==3.0.3
idna==3.6
packaging==23.2
pychrome==0.2.4
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
requests==2.31.0
urllib3==2.2.0
websocket==0.2.1
websocket-client==1.7.0
zope.event==5.0
zope.interface==6.1
